# api-gateway/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "QuickServe API",
    "description": "API pour l'application QuickServe de transport et services",
    "version": "1.0.0",
    "contact": {
      "name": "Équipe QuickServe",
      "email": "support@quickserve.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Serveur de développement local"
    },
    {
      "url": "https://api.quickserve.com/api",
      "description": "Serveur de production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Message d'erreur"
          }
        }
      },
...
      "Location": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "float",
            "description": "Latitude"
          },
          "longitude": {
            "type": "number",
            "format": "float",
            "description": "Longitude"
          },
          "accuracy": {
            "type": "number",
            "format": "float",
            "description": "Précision en mètres"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Date de la dernière mise à jour"
          }
        }
      },
...
      "PaymentIntent": {
        "type": "object",
        "properties": {
          "client_secret": {
            "type": "string",
            "description": "Client secret pour l'intention de paiement Stripe"
          }
        }
      }
    }
  },
  "paths": {
...
    "/providers/{id}/location": {
      "get": {
        "tags": [
          "Géolocalisation"
        ],
        "summary": "Position d'un prestataire",
        "description": "Récupérer la dernière position connue d'un prestataire",
        "operationId": "getProviderLocation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du prestataire",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Position récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location": {
                      "$ref": "#/components/schemas/Location"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Position non trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/providers/nearby": {
      "get": {
        "tags": [
          "Géolocalisation"
        ],
        "summary": "Prestataires à proximité",
        "description": "Trouver les prestataires à proximité d'une position donnée",
        "operationId": "getNearbyProviders",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "latitude",
            "in": "query",
            "description": "Latitude de la position",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "longitude",
            "in": "query",
            "description": "Longitude de la position",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "description": "Rayon de recherche en kilomètres",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des prestataires à proximité récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "providers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "provider_id": {
                            "type": "integer",
                            "format": "int64",
                            "description": "Identifiant du prestataire"
                          },
                          "latitude": {
                            "type": "number",
                            "format": "float",
                            "description": "Latitude de la position"
                          },
                          "longitude": {
                            "type": "number",
                            "format": "float",
                            "description": "Longitude de la position"
                          },
                          "distance": {
                            "type": "number",
                            "format": "float",
                            "description": "Distance en kilomètres"
                          },
                          "updated_at": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Date de la dernière mise à jour de la position"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Coordonnées invalides",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/services/{id}/location-history": {
      "get": {
        "tags": [
          "Géolocalisation"
        ],
        "summary": "Historique des positions",
        "description": "Récupérer l'historique des positions pour un service",
        "operationId": "getServiceLocationHistory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifiant du service",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historique récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "history": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "latitude": {
                            "type": "number",
                            "format": "float",
                            "description": "Latitude"
                          },
                          "longitude": {
                            "type": "number",
                            "format": "float",
                            "description": "Longitude"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Date d'enregistrement de la position"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Service non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/payments/payment-intents": {
      "post": {
        "tags": [
          "Paiement"
        ],
        "summary": "Créer une intention de paiement",
        "description": "Créer une intention de paiement pour un service",
        "operationId": "createPaymentIntent",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "service_id"
                ],
                "properties": {
                  "service_id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Identifiant du service à payer"
                  },
                  "payment_method_id": {
                    "type": "integer",
                    "format": "int64",
                    "description": "Identifiant de la méthode de paiement (optionnel)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intention de paiement créée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentIntent"
                }
              }
            }
          },
          "400": {
            "description": "Données invalides ou service déjà payé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": null,
              "schema": {
                "$ref": "#/components/schemas/Error"
              }
            }
          },
          "403": {
            "description": "Accès non autorisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/payments/confirm": {
      "post": {
        "tags": [
          "Paiement"
        ],
        "summary": "Confirmer un paiement",
        "description": "Confirmer un paiement après succès de l'opération chez Stripe",
        "operationId": "confirmPayment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "payment_intent_id"
                ],
                "properties": {
                  "payment_intent_id": {
                    "type": "string",
                    "description": "Identifiant de l'intention de paiement Stripe"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paiement confirmé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Paiement confirmé avec succès"
                    },
                    "payment_id": {
                      "type": "integer",
                      "format": "int64",
                      "description": "Identifiant du paiement"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Paiement non validé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Accès non autorisé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Paiement non trouvé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/payments/history": {
      "get": {
        "tags": [
          "Paiement"
        ],
        "summary": "Historique des paiements",
        "description": "Récupérer l'historique des paiements du client",
        "operationId": "getPaymentHistory",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Historique des paiements récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "format": "int64",
                            "description": "Identifiant du paiement"
                          },
                          "service_id": {
                            "type": "integer",
                            "format": "int64",
                            "description": "Identifiant du service associé"
                          },
                          "service_title": {
                            "type": "string",
                            "description": "Titre du service"
                          },
                          "amount": {
                            "type": "number",
                            "format": "float",
                            "description": "Montant du paiement"
                          },
                          "status": {
                            "type": "string",
                            "enum": [
                              "pending",
                              "completed",
                              "failed",
                              "cancelled"
                            ],
                            "description": "Statut du paiement"
                          },
                          "payment_method": {
                            "type": "string",
                            "description": "Méthode de paiement utilisée"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Date de création du paiement"
                          },
                          "processed_at": {
                            "type": "string",
                            "format": "date-time",
                            "description": "Date de traitement du paiement"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}